factor out settings, renderable, and instrument
