+++ shared/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod helpers;
 mod settings;
 
 pub use settings::{default_settings, get_test_settings, Settings};
