+++ src/generation/mod.rs
@@ -1,6 +1,5 @@
 pub mod json;
 pub mod parsed_to_render;
-
 pub mod test;
 
 pub use self::{
