+++ src/generation/parsed_to_render.rs
@@ -1,7 +1,5 @@
 use crate::{
     generation::{to_csv, to_json},
-    instrument::{Basis, Oscillator, StereoWaveform},
-    renderable::{nf_to_vec_renderable, renderables_to_render_voices, RenderOp, Renderable},
     ui::printed,
     write::{write_composition_to_mp3, write_composition_to_wav},
 };
@@ -14,6 +12,10 @@ use std::path::Path;
 use std::sync::{Arc, Mutex};
 use weresocool_ast::{Defs, NormalForm, Term};
 use weresocool_error::{Error, IdError};
+use weresocool_instrument::renderable::{
+    nf_to_vec_renderable, renderables_to_render_voices, RenderOp, Renderable,
+};
+use weresocool_instrument::{Basis, Oscillator, StereoWaveform};
 use weresocool_parser::ParsedComposition;
 use weresocool_shared::{default_settings, Settings};
