--- src/instrument/mod.rs
@@ -1,19 +1,0 @@
-mod asr;
-mod frequency;
-mod gain;
-mod loudness;
-pub mod oscillator;
-mod sample;
-pub mod stereo_waveform;
-pub mod voice;
-
-#[cfg(test)]
-mod asr_test;
-#[cfg(test)]
-#[allow(clippy::unreadable_literal)]
-mod test;
-
-pub use self::{
-    oscillator::{Basis, Oscillator},
-    stereo_waveform::{Normalize, StereoWaveform},
-};
