+++ src/lib.rs
@@ -8,7 +8,6 @@ pub mod analyze;
 pub mod examples;
 pub mod generation;
 pub mod helpers;
-pub mod instrument;
 pub mod interpretable;
 pub mod manager;
 pub mod portaudio;
