+++ src/portaudio/real_time.rs
@@ -1,11 +1,9 @@
-use crate::{
-    generation::parsed_to_render::sum_all_waveforms, instrument::StereoWaveform,
-    renderable::RenderVoice, write::write_output_buffer,
-};
+use crate::{generation::parsed_to_render::sum_all_waveforms, write::write_output_buffer};
 use portaudio as pa;
 use rayon::prelude::*;
-
 use weresocool_error::Error;
+use weresocool_instrument::RenderVoice;
+use weresocool_instrument::StereoWaveform;
 use weresocool_shared::{default_settings, Settings};
 
 const SETTINGS: Settings = default_settings();
